увеличение интервала номера до 100

+++ src/test/java/main/FiboTest.java
@@ -3,13 +3,15 @@ package main;
 import org.junit.Assert;
 import org.junit.Test;
 
+import java.math.BigInteger;
+
 public class FiboTest {
 
     @Test
     public void calculationResult() {
         int in = 6;
         int expected = 8;
-        int out = new Fibo().calculationResult(in);
+        BigInteger out = new Fibo().calculationResult(in);
         Assert.assertEquals(expected, out);
     }
 }
